saving to postgres does not need to be an api
